Abstract Class is done
I delete the addTocollectedCards function because collectCards function is enoguh.

I change the name of the playCard function to playCardTo and gave an implemetation to this and ı also added a paramter to this function.

ı gave an implemetation to addToCurrentCards function and collectCard function and ı gave parameters to them.

In user interface ı rearranged the interfaces of the functions.